Improve matching of python sections in mmm-mode
First: there exist python sections with the simple format:

   python () {
      ...
   }

Deal with this by making the name (function name?) after the python
keyword optional. Now to avoid clashes with the bitbake-shell class,
we need to add a front-verify function there to disambiguate.

Second: there seem to exist inline python definitions that just start
with the "def" keyword directly. We make a separate mmm class for this
that matches from def until the first non-empty non-indented line,
which seems to be in line with how the bbclasses I've seen are to be
read.

Also improve the original regexes to use [[:space:]] in many
places where only one kind of whitespace was previously accepted.